+++ src/models/MiniReel.js
@@ -48,10 +48,11 @@ function getCardType(card) {
     }
 }
 
-function initialize(whitelist, { experience, standalone, profile }) {
+function initialize(whitelist, { experience, standalone, interstitial, profile }) { // jshint ignore:line
     const deck = filter(experience.data.deck, card => whitelist.indexOf(getCardType(card)) > -1);
 
     this.standalone = standalone;
+    this.interstitial = interstitial;
     this.id = experience.id;
     this.title = experience.data.title;
     this.branding = experience.data.branding;
@@ -131,6 +132,7 @@ export default class MiniReel extends EventEmitter {
         super(...arguments);
 
         this.standalone = null;
+        this.interstitial = null;
 
         this.id = null;
         this.title = null;
@@ -150,6 +152,7 @@ export default class MiniReel extends EventEmitter {
         this.currentIndex = -1;
         this.currentCard = null;
         this.skippable = true;
+        this.closeable = true;
 
         _(this).ready = false;
         _(this).cardsShown = 0;
@@ -185,6 +188,19 @@ export default class MiniReel extends EventEmitter {
 
         this.on('launch', () => cinema6.getSession().then(session => session.ping('open')));
         this.on('close', () => cinema6.getSession().then(session => session.ping('close')));
+
+        this.on('becameUnskippable', () => {
+            if (this.interstitial) {
+                this.closeable = false;
+                this.emit('becameUncloseable');
+            }
+        });
+        this.on('becameSkippable', () => {
+            if (!this.closeable) {
+                this.closeable = true;
+                this.emit('becameCloseable');
+            }
+        });
 
         dispatcher.addClient(ADTECHHandler);
         dispatcher.addClient(PostMessageHandler, window.parent.postMessage);